Add files via upload
This application is used to analyze the thermoluminescence glow curve and evaluate the trapping parameters using an analytical equation derived from the one trap - one recombination model. Running this application requires only to set up MATLAB and does not require any knowledge of writing codes. To run the applications, simply follow these steps:
Set the path in the MATLAB main command window to the folder where the file was downloaded.
Right-click on the file, then install.